Rename interfaces/ to types/, add types/index.ts
