+++ beaconChain/types/state.ts
